requirements.txt
asgiref==3.5.2
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
dj-database-url==1.0.0
Django==3.2.15
django-cors-headers==3.13.0
django-heroku==0.3.1
django-ninja==0.19.1
gunicorn==20.1.0
heroku==0.1.4
idna==3.4
Pillow==9.2.0
psycopg2==2.9.3
pycparser==2.21
pydantic==1.10.2
python-dateutil==1.5
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
typing-extensions==4.3.0
urllib3==1.26.12
whitenoise==6.2.0
